move setNumHbftClients to HbftJavaScenario
